move http package down into tr069 package
