+++ tr069/src/main/java/com/github/freeacs/controllers/Tr069Controller.java
@@ -4,9 +4,9 @@ import com.github.freeacs.dbi.ACS;
 import com.github.freeacs.dbi.ACSUnit;
 import com.github.freeacs.dbi.DBI;
 import com.github.freeacs.dbi.Unit;
-import com.github.freeacs.http.HTTPRequestData;
-import com.github.freeacs.http.HTTPRequestResponseData;
-import com.github.freeacs.http.HTTPResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPResponseData;
 import com.github.freeacs.tr069.Properties;
 import com.github.freeacs.tr069.SessionData;
 import com.github.freeacs.tr069.SessionLogging;
