+++ tr069/src/main/java/com/github/freeacs/tr069/ParameterKey.java
@@ -6,7 +6,7 @@ import com.github.freeacs.dbi.JobParameter;
 import com.github.freeacs.dbi.UnittypeParameter;
 import com.github.freeacs.dbi.UnittypeParameters;
 import com.github.freeacs.dbi.util.SystemParameters;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.xml.ParameterValueStruct;
 import lombok.Data;
