+++ tr069/src/main/java/com/github/freeacs/tr069/SessionLogging.java
@@ -8,9 +8,9 @@ import com.github.freeacs.dbi.util.ProvisioningMessage.ProvOutput;
 import com.github.freeacs.dbi.util.ProvisioningMessage.ProvStatus;
 import com.github.freeacs.dbi.util.ProvisioningMode;
 import com.github.freeacs.dbi.util.SyslogClient;
-import com.github.freeacs.http.HTTPRequestData;
-import com.github.freeacs.http.HTTPRequestResponseData;
-import com.github.freeacs.http.HTTPResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPResponseData;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
 import com.github.freeacs.tr069.xml.ParameterValueStruct;
 import java.util.List;
