+++ tr069/src/main/java/com/github/freeacs/tr069/http/HTTPRequestResponseData.java
@@ -1,4 +1,4 @@
-package com.github.freeacs.http;
+package com.github.freeacs.tr069.http;
 
 import com.github.freeacs.tr069.SessionData;
 import com.github.freeacs.tr069.base.BaseCache;
