+++ tr069/src/main/java/com/github/freeacs/tr069/methods/decision/GetRPCMethodsDecisionStrategy.java
@@ -1,6 +1,6 @@
 package com.github.freeacs.tr069.methods.decision;
 
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
 
 public class GetRPCMethodsDecisionStrategy implements DecisionStrategy {
