+++ tr069/src/main/java/com/github/freeacs/tr069/methods/decision/SetParameterValuesDecisionStrategy.java
@@ -5,7 +5,7 @@ import com.github.freeacs.tr069.base.Log;
 import com.github.freeacs.tr069.base.UnitJob;
 import com.github.freeacs.dbi.UnitJobStatus;
 import com.github.freeacs.dbi.util.ProvisioningMode;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.Properties;
 import com.github.freeacs.tr069.SessionData;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
