+++ tr069/src/main/java/com/github/freeacs/tr069/methods/decision/TransferCompleteDecisionStrategy.java
@@ -1,7 +1,7 @@
 package com.github.freeacs.tr069.methods.decision;
 
 import com.github.freeacs.tr069.base.Log;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
 import com.github.freeacs.tr069.xml.Fault;
