+++ tr069/src/main/java/com/github/freeacs/tr069/methods/request/RebootRequestProcessStrategy.java
@@ -1,6 +1,6 @@
 package com.github.freeacs.tr069.methods.request;
 
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
 import com.github.freeacs.tr069.xml.Parser;
