+++ tr069/src/main/java/com/github/freeacs/tr069/methods/request/RequestProcessStrategy.java
@@ -2,7 +2,7 @@ package com.github.freeacs.tr069.methods.request;
 
 import com.github.freeacs.dbi.DBI;
 import com.github.freeacs.tr069.base.Log;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.Properties;
 import com.github.freeacs.tr069.methods.ProvisioningMethod;
