+++ tr069/src/main/java/com/github/freeacs/tr069/methods/response/DownloadResponseCreateStrategy.java
@@ -3,7 +3,7 @@ package com.github.freeacs.tr069.methods.response;
 import com.github.freeacs.dbi.FileType;
 import com.github.freeacs.dbi.util.ProvisioningMessage;
 import com.github.freeacs.dbi.util.SystemParameters;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.Properties;
 import com.github.freeacs.tr069.SessionData;
 import com.github.freeacs.tr069.xml.Body;
