+++ tr069/src/main/java/com/github/freeacs/tr069/methods/response/GetParameterValuesResponseCreateStrategy.java
@@ -3,7 +3,7 @@ package com.github.freeacs.tr069.methods.response;
 import com.github.freeacs.tr069.base.Log;
 import com.github.freeacs.dbi.UnittypeParameters;
 import com.github.freeacs.dbi.util.ProvisioningMode;
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.CPEParameters;
 import com.github.freeacs.tr069.Properties;
 import com.github.freeacs.tr069.SessionData;
