+++ tr069/src/main/java/com/github/freeacs/tr069/methods/response/TransferCompleteResponseCreateStrategy.java
@@ -1,6 +1,6 @@
 package com.github.freeacs.tr069.methods.response;
 
-import com.github.freeacs.http.HTTPRequestResponseData;
+import com.github.freeacs.tr069.http.HTTPRequestResponseData;
 import com.github.freeacs.tr069.xml.Body;
 import com.github.freeacs.tr069.xml.Header;
 import com.github.freeacs.tr069.xml.Response;
